Move a page creation function
